requirements.txt
pytube==12.1.2
streamlit
openai-whisper
moviepy